+++ src/main/java/com/minorjava/week3/controller/PersonController.java
@@ -3,13 +3,19 @@ package com.minorjava.week3.controller;
 import com.minorjava.week3.model.Person;
 import com.minorjava.week3.service.IPersonService;
 import com.minorjava.week3.util.ErrorMessage;
+import com.minorjava.week3.util.FieldErrorMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import javax.xml.bind.ValidationException;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 public class PersonController {
@@ -18,15 +24,17 @@ public class PersonController {
     IPersonService personService;
 
     @PostMapping("/person")
-    Person create(@RequestBody Person person) throws ValidationException {
-        if(person.getId()==0 && person.getFirstName()!=null && person.getLastName()!=null)
-        {
+    Person create(@Valid @RequestBody Person person){
             return personService.save(person);
-        } else
-        {
-            throw new ValidationException("Person cannot be created");
-        }
+           }
+   @ResponseStatus(HttpStatus.BAD_GATEWAY)
+   @ExceptionHandler(MethodArgumentNotValidException.class)
+   List<FieldErrorMessage> exceptionHandler(MethodArgumentNotValidException e){
+        List<FieldError> fieldErrors = e.getBindingResult().getFieldErrors();
+       List<FieldErrorMessage> fieldErrorMessages = fieldErrors.stream().map(fieldError -> new FieldErrorMessage(fieldError.getField(), fieldError.getDefaultMessage())).collect(Collectors.toList());
+    return fieldErrorMessages;
     }
+
 
     @GetMapping("/person")
     Iterable<Person> read() {
@@ -42,8 +50,12 @@ public class PersonController {
         }
     }
     @DeleteMapping("/person/{id}")
-    void delete(@PathVariable Integer id){
-        personService.delete(id);
+    void delete(@PathVariable Integer id) throws ValidationException {
+        if(personService.findById(id).isPresent()){
+            personService.delete(id);
+        } else {
+            throw new ValidationException("Person cannot be deleted");
+        }
     }
     @GetMapping("/person/{id}")
     Optional<Person> findById(@PathVariable Integer id){
